Adjusts HOME

+++ lib/models/home_sections/home_manager.dart
@@ -1,4 +1,4 @@
-import 'package:brn_ecommerce/models/sections_home/section.dart';
+import 'package:brn_ecommerce/models/home_sections/section.dart';
 import 'package:brn_ecommerce/services/development_monitoring/firebase_performance.dart';
 import 'package:brn_ecommerce/services/development_monitoring/monitoring_logger.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
@@ -6,9 +6,9 @@ import 'package:flutter/foundation.dart';
 
 /// # HomeManager (Folder: models/views)
 ///
-/// A class responsible for managing home page sections_home and their editing state.
+/// A class responsible for managing home page home_sections and their editing state.
 ///
-/// This class handles loading sections_home, adding, removing, moving sections_home, entering editing mode,
+/// This class handles loading home_sections, adding, removing, moving home_sections, entering editing mode,
 /// saving changes, and discarding edits for the home page.
 class HomeManager extends ChangeNotifier {
   // proprieties
@@ -21,7 +21,7 @@ class HomeManager extends ChangeNotifier {
 
   // Constructor
 
-  /// Initializes a [HomeManager] instance and loads home page sections_home.
+  /// Initializes a [HomeManager] instance and loads home page home_sections.
   HomeManager() {
     _loadSections();
   }
@@ -31,9 +31,9 @@ class HomeManager extends ChangeNotifier {
   List<Section> get sections => editing ? _editingSections : _sections;
 
   // Methods
-  /// Loads home page sections_home from Firestore.
+  /// Loads home page home_sections from Firestore.
   Future<void> _loadSections() async {
-    PerformanceMonitoring().startTrace('load-sections_home', shouldStart: true);
+    PerformanceMonitoring().startTrace('load-home_sections', shouldStart: true);
     if (kDebugMode) MonitoringLogger().logInfo('Starting listen Sections');
 
     firestore.collection("home").orderBy("position").snapshots().listen((snapshot) {
@@ -43,10 +43,11 @@ class HomeManager extends ChangeNotifier {
       }
 
       _ensureLocalBestSellingSection();
+      _ensureLocalRecentlyAddedSection();
       notifyListeners();
     });
 
-    PerformanceMonitoring().stopTrace('load-sections_home');
+    PerformanceMonitoring().stopTrace('load-home_sections');
     if (kDebugMode) MonitoringLogger().logInfo('Ending listen Sections');
   }
 
@@ -61,20 +62,31 @@ class HomeManager extends ChangeNotifier {
     }
   }
 
-  /// Adds a new section to the list of editing sections_home.
+  void _ensureLocalRecentlyAddedSection() {
+    final exists = _sections.any((s) => s.type == 'RecentlyAdded');
+    if (!exists) {
+      final indexToInsert = _sections.indexWhere((s) => s.type == 'List');
+      _sections.insert(
+        indexToInsert >= 0 ? indexToInsert + 4 : 0,
+        Section(type: 'RecentlyAdded', name: 'Adicionados Recentemente'),
+      );
+    }
+  }
+
+  /// Adds a new section to the list of editing home_sections.
   void addSection(Section section) {
     _editingSections.add(section);
     notifyListeners();
   }
 
-  /// Removes a section from the list of editing sections_home.
+  /// Removes a section from the list of editing home_sections.
   void removeSection(Section section) {
-    if (section.type == "BestSelling") return;
+    if (section.type == "BestSelling" || section.type == "RecentlyAdded") return;
     _editingSections.remove(section);
     notifyListeners();
   }
 
-  /// Moves a section up in the list of editing sections_home.
+  /// Moves a section up in the list of editing home_sections.
   void moveSectionUp(Section section) {
     final idx = _editingSections.indexOf(section);
     if (idx > 0) {
@@ -83,7 +95,7 @@ class HomeManager extends ChangeNotifier {
     }
   }
 
-  /// Moves a section down in the list of editing sections_home.
+  /// Moves a section down in the list of editing home_sections.
   void moveSectionDown(Section section) {
     final idx = _editingSections.indexOf(section);
     if (idx < _editingSections.length - 1) {
@@ -92,7 +104,7 @@ class HomeManager extends ChangeNotifier {
     }
   }
 
-  /// Swaps the positions of two sections_home in the list of editing sections_home.
+  /// Swaps the positions of two home_sections in the list of editing home_sections.
   void _swapSections(Section s1, Section s2) {
     final idx1 = _editingSections.indexOf(s1);
     final idx2 = _editingSections.indexOf(s2);
@@ -103,24 +115,12 @@ class HomeManager extends ChangeNotifier {
     }
   }
 
-  /// Enters editing mode by cloning current sections_home for editing.
+  /// Enters editing mode by cloning current home_sections for editing.
   void enterEditing() {
     editing = true;
 
     // Clone all existing sections
     _editingSections = _sections.map((s) => s.clone()).toList();
-
-    // Ensures that BestSelling exists in the edit list
-    final exists = _editingSections.any((s) => s.type == 'BestSelling');
-    if (!exists) {
-      // Place after the first section of type 'List' or at the beginning
-      final indexToInsert = _editingSections.indexWhere((s) => s.type == 'List');
-      _editingSections.insert(
-        indexToInsert >= 0 ? indexToInsert + 1 : 0,
-        Section(type: 'BestSelling', name: 'Mais Vendidos'),
-      );
-    }
-
     notifyListeners();
   }
 
@@ -140,14 +140,14 @@ class HomeManager extends ChangeNotifier {
 
     int position = 0;
     for (final section in _editingSections) {
-      if (section.type != "BestSelling") {
+      if (section.type != "BestSelling" || section.type != "RecentlyAdded") {
         await section.saveSection(position);
       }
       position++;
     }
 
     for (final section in List.from(_sections)) {
-      if (section.type == "BestSelling") continue;
+      if (section.type == "BestSelling" || section.type == "RecentlyAdded") continue;
       if (!_editingSections.any((e) => e.id == section.id)) {
         await section.delete();
       }
@@ -161,6 +161,7 @@ class HomeManager extends ChangeNotifier {
   /// Discards changes made in editing mode.
   void discardEditing() {
     editing = false;
+    _editingSections.clear();
     notifyListeners();
   }
 }
